+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/procedures/Payment.java
@@ -25,7 +25,6 @@ import com.oltpbenchmark.benchmarks.tpcc.TPCCWorker;
 import com.oltpbenchmark.benchmarks.tpcc.pojo.Customer;
 import com.oltpbenchmark.benchmarks.tpcc.pojo.District;
 import com.oltpbenchmark.benchmarks.tpcc.pojo.Warehouse;
-import java.math.BigDecimal;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.Random;
@@ -309,7 +308,7 @@ public class Payment extends TPCCProcedure {
 
   private void updateWarehouse(Connection conn, int w_id, float paymentAmount) throws SQLException {
     try (PreparedStatement payUpdateWhse = this.getPreparedStatement(conn, payUpdateWhseSQL)) {
-      payUpdateWhse.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+      payUpdateWhse.setDouble(1, paymentAmount);
       payUpdateWhse.setInt(2, w_id);
       // MySQL reports deadlocks due to lock upgrades:
       // t1: read w_id = x; t2: update w_id = x; t1 update w_id = x
@@ -378,7 +377,7 @@ public class Payment extends TPCCProcedure {
   private void updateDistrict(Connection conn, int w_id, int districtID, float paymentAmount)
       throws SQLException {
     try (PreparedStatement payUpdateDist = this.getPreparedStatement(conn, payUpdateDistSQL)) {
-      payUpdateDist.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+      payUpdateDist.setDouble(1, paymentAmount);
       payUpdateDist.setInt(2, w_id);
       payUpdateDist.setInt(3, districtID);
